requirements.txt
# Requirements for the project
confluent-kafka==2.2.0
kafka-python>=2.0.2
pymongo>=4.0
requests>=2.28.0
tabulate